RISKY: relocates checkbox-tree to config lib
