+++ libs/spirit-islander/ui/src/index.ts
@@ -1,7 +1,5 @@
 export * from './lib/board-emblem/board-emblem.component';
 export * from './lib/card/card.component';
-export * from './lib/checkbox-tree/checkbox-tree.component';
-export * from './lib/checkbox-tree/create-tree';
 export * from './lib/difficulty-emblem/difficulty-emblem.component';
 export * from './lib/expansion-emblem/expansion-emblem.component';
 export * from './lib/card-group/card-group.component';
